move error simulation into its own directory
