+++ src/amo/routes.js
@@ -2,7 +2,8 @@ import config from 'config';
 import React from 'react';
 import { IndexRoute, Route } from 'react-router';
 
-import SimulateError from 'core/containers/SimulateError';
+import SimulateSyncError from
+  'core/containers/error-simulation/SimulateSyncError';
 import HandleLogin from 'core/containers/HandleLogin';
 
 import AddonReviewList from './components/AddonReviewList';
@@ -34,7 +35,7 @@ export default (
     <Route path="404/" component={NotFound} />
     <Route path="500/"
       component={config.get('isDevelopment') ? ServerError : NotFound} />
-    <Route path="simulate-error/" component={SimulateError} />
+    <Route path="simulate-sync-error/" component={SimulateSyncError} />
     <Route path=":visibleAddonType/" component={LandingPage} />
     <Route path="*" component={NotFound} />
   </Route>
