+++ src/core/containers/error-simulation/SimulateSyncError.js
