+++ src/disco/routes.js
@@ -4,7 +4,8 @@ import { Router, Route } from 'react-router';
 
 import GenericError from 'core/components/ErrorPage/GenericError';
 import NotFound from 'core/components/ErrorPage/NotFound';
-import SimulateError from 'core/containers/SimulateError';
+import SimulateSyncError from
+  'core/containers/error-simulation/SimulateSyncError';
 
 import App from './containers/App';
 import DiscoPane from './containers/DiscoPane';
@@ -19,7 +20,10 @@ export default (
     <Route path="/:lang/firefox/404" component={NotFound} />
     <Route path="/:lang/firefox/500"
       component={config.get('isDevelopment') ? GenericError : NotFound} />
-    <Route path="/:lang/firefox/simulate-error/" component={SimulateError} />
+    <Route
+      path="/:lang/firefox/simulate-sync-error/"
+      component={SimulateSyncError}
+    />
     <Route path="*" component={NotFound} />
   </Router>
 );
